+++ paddle_billing/Entities/Reports/__init__.py
@@ -1,4 +1,5 @@
 from paddle_billing.Entities.Reports.AdjustmentsReportType import AdjustmentsReportType
+from paddle_billing.Entities.Reports.BalanceMovementType import BalanceMovementType
 from paddle_billing.Entities.Reports.BalanceReportType import BalanceReportType
 from paddle_billing.Entities.Reports.CheckoutsReportType import CheckoutsReportType
 from paddle_billing.Entities.Reports.DiscountsReportType import DiscountsReportType