+++ paddle_billing/Notifications/Entities/Reports/ReportFilterName.py
@@ -3,6 +3,8 @@ from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
 
 class ReportFilterName(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
     Action: "ReportFilterName" = "action"
+    BalanceMovementDate: "ReportFilterName" = "balance_movement_date"
+    BalanceMovementType: "ReportFilterName" = "balance_movement_type"
     CheckoutCreatedAt: "ReportFilterName" = "checkout_created_at"
     CollectionMode: "ReportFilterName" = "collection_mode"
     CurrencyCode: "ReportFilterName" = "currency_code"