+++ tests/Functional/Resources/Reports/test_ReportsClient.py
@@ -7,6 +7,7 @@ from paddle_billing.Entities.Report import Report
 from paddle_billing.Entities.ReportCSV import ReportCSV
 from paddle_billing.Entities.Reports import (
     AdjustmentsReportType,
+    BalanceMovementType,
     BalanceReportType,
     CheckoutsReportType,
     DiscountsReportType,
@@ -24,6 +25,8 @@ from paddle_billing.Entities.DateTime import DateTime
 from paddle_billing.Resources.Reports.Operations.Filters import (
     AdjustmentActionFilter,
     AdjustmentStatusFilter,
+    BalanceMovementDateFilter,
+    BalanceMovementTypeFilter,
     CheckoutCreatedAtFilter,
     CollectionModeFilter,
     CurrencyCodeFilter,
@@ -196,6 +199,24 @@ class TestReportsClient:
                 ReadsFixtures.read_raw_json_fixture("request/create_checkouts_full"),
                 CheckoutsReportType,
             ),
+            (
+                CreatePayoutReconciliationReport(
+                    type=PayoutReconciliationReportType.PayoutReconciliation,
+                    filters=[
+                        BalanceMovementTypeFilter(
+                            [
+                                BalanceMovementType.Sale,
+                                BalanceMovementType.Chargeback,
+                                BalanceMovementType.Credit,
+                            ]
+                        ),
+                        BalanceMovementDateFilter.gte(DateTime("2023-12-30")),
+                        BalanceMovementDateFilter.lt(DateTime("2024-12-30T12:30:01.123456Z")),
+                    ],
+                ),
+                ReadsFixtures.read_raw_json_fixture("request/create_payout_reconciliation_balance_movement"),
+                PayoutReconciliationReportType,
+            ),
         ],
         ids=[
             "Create transaction report with basic data",
@@ -209,6 +230,7 @@ class TestReportsClient:
             "Create payout reconciliation report with filters",
             "Create checkouts report with basic data",
             "Create checkouts report with filters",
+            "Create payout reconciliation report with balance movement filters",
         ],
     )
     def test_create_report_uses_expected_payload(